+++ apps/meteor/client/methods/index.ts
@@ -1,5 +1,1 @@
-import './hideRoom';
 import './openRoom';
-import './pinMessage';
-import './unpinMessage';
-import './updateMessage';